Install Pelican and plugin with Nix Flake
Pelican is a static site generator written in Python.

Content for this site will be written in Pandoc compatible Markdown,
which supports math, citations and more structured metadata. A plugin is
available, which will let Pelican use Pandoc to parse/render the
Markdown files.

Pelican itself can be installed from nixpkgs, but Pelican plugins are
not available there. The following workaround is used:

-   Nix is used install the target Python version.
-   The package manager `uv` will be used to manage the Python
    dependencies. This package manager is fast enough to run every time
    a shell is started, and it is also deterministic through the use of
    lock files and the `sync` command.
-   The list of required dependencies is located in `requirements.in`,
    which is 'locked' to create `requirements.txt`.
-   When starting a development shell -- e.g. `nix develop` -- `uv` will
    ensure a virtual environment exists with only the packages
    specified, and this virtual environment will also be activated.

Pandoc is also installed through nixpkgs.